% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureset.R
\name{commit_featureset}
\alias{commit_featureset}
\title{Dkube Featureset}
\usage{
commit_featureset(
  name = NULL,
  df = NULL,
  path = NULL,
  filepath = "/tmp/metadata.yaml"
)
}
\arguments{
\item{name}{featureset name}

\item{df}{dataframe}

\item{path}{output mount path}

\item{filepath}{parquet file name}
}
\description{
Dkube Featureset
}
