% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repos.R
\name{DkubeFeatureSet}
\alias{DkubeFeatureSet}
\title{Dkube Featureset}
\usage{
DkubeFeatureSet(name = NULL, description = NULL)
}
\arguments{
\item{name}{repo/resource name}

\item{description}{additional details for the repo}

\item{user}{username}
}
\description{
Dkube Featureset
}
