% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureset.R
\name{featureset_commit}
\alias{featureset_commit}
\title{Dkube Featureset}
\usage{
featureset_commit(
  name = NULL,
  df = NULL,
  path = NULL,
  filepath = "/tmp/metadata.yaml"
)
}
\arguments{
\item{name}{featureset name}

\item{df}{dataframe}

\item{path}{output mount path}

\item{filepath}{parquet file name}
}
\description{
Dkube Featureset
}
