% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/passwd.R
\name{get_redshift_password}
\alias{get_redshift_password}
\title{Dkube get password}
\usage{
get_redshift_password(user, db)
}
\arguments{
\item{user}{redshift user}

\item{db}{DB name}
}
\description{
Dkube get password
}
