% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureset.R
\name{read_features}
\alias{read_features}
\title{Dkube Featureset}
\usage{
read_features(name = NULL, filename = "featureset.parquet", path = NULL)
}
\arguments{
\item{filename}{parquet file name}

\item{path}{input mount path}

\item{df}{dataframe}
}
\description{
Dkube Featureset
}
