% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureset.R
\name{write_feature}
\alias{write_feature}
\title{Dkube Featureset}
\usage{
write_feature(df, filename = "featureset.parquet", path = NULL)
}
\arguments{
\item{df}{dataframe}

\item{filename}{parquet file name}

\item{path}{output mount path}
}
\description{
Dkube Featureset
}
