% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureset.R
\name{write_metadata}
\alias{write_metadata}
\title{Dkube Featureset}
\usage{
write_metadata(df, filepath = NULL)
}
\arguments{
\item{df}{dataframe}

\item{filename}{yaml file name to write metadata}
}
\description{
Dkube Featureset
}
